+++ _default/view/home.php
@@ -3,7 +3,7 @@ if (!defined('MAIN')) {
     require $_SERVER['ROOT_PATH'] . $_SERVER['ERROR_PATH'];
 }
 ?>
-<div class="main">
+<div class="main">   
 
     <img src="<?php echo $this->APP_URL; ?>ico1.png"/>